[design] Landing: 텍스트 애니메이션 초안
