+++ src/app/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import LandingTopTest from '@/components/Landing/LandingTopTest';
+import LandingTopSection from '@/components/Landing/LandingTopSection';
 
 export default function LandingPage() {
   return (
     <div className="flex h-screen w-full flex-col">
-      <LandingTopTest />
+      <LandingTopSection />
     </div>
   );
 }
